Fix cpp compiler warnings (#797)
### Changelog
None

### Docs
None

### Description
- include websocketpp & nlohman_json as system libs to suppress compiler
warnings
- fix conversion warning
- fix declaration hides class member warning

+++ cpp/foxglove-websocket/include/foxglove/websocket/websocket_server.hpp
@@ -1536,7 +1536,7 @@ inline void Server<ServerConfiguration>::sendFetchAssetResponse(
   const uint8_t status = static_cast<uint8_t>(response.status);
   message->append_payload(&status, 1);
 
-  foxglove::WriteUint32LE(uint32Data.data(), response.errorMessage.size());
+  foxglove::WriteUint32LE(uint32Data.data(), static_cast<uint32_t>(response.errorMessage.size()));
   message->append_payload(uint32Data.data(), uint32Data.size());
   message->append_payload(response.errorMessage.data(), errMsgSize);
